wiring logic of kitchen and cleaning the request to dispatch drivers

+++ orderserver/src/main/java/com/acabra/orderfullfilment/orderserver/dto/OrderMapper.java
@@ -3,7 +3,7 @@ package com.acabra.orderfullfilment.orderserver.dto;
 import com.acabra.orderfullfilment.orderserver.model.DeliveryOrder;
 
 public class OrderMapper {
-    public DeliveryOrder fromDeliveryOrderRequest(DeliveryOrderRequest deliveryOrderRequest, int arrivalTime) {
+    public DeliveryOrder fromDeliveryOrderRequest(DeliveryOrderRequest deliveryOrderRequest, long arrivalTime) {
         return new DeliveryOrder(deliveryOrderRequest.id,
                 deliveryOrderRequest.name,
                 deliveryOrderRequest.prepTime,
